pyproject.toml
[project]
name = "chatbot"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "alembic (>=1.18.4,<2.0.0)",
    "sqlalchemy[asyncio] (>=2.0.49,<3.0.0)",
    "sqlmodel (>=0.0.38,<0.0.39)",
    "pytest (>=9.0.2,<10.0.0)",
    "pytest-asyncio (>=1.3.0,<2.0.0)",
    "pydantic-settings (>=2.13.1,<3.0.0)",
    "psycopg2-binary (>=2.9.11,<3.0.0)",
    "asyncpg (>=0.31.0,<0.32.0)",
    "starkbank (>=2.27.0,<3.0.0)",
    "email-validator (>=2.3.0,<3.0.0)"
]


[tool.mypy]
python_version = "3.14"
ignore_missing_imports = true
disallow_untyped_defs = true
disallow_untyped_calls = true
explicit_package_bases = true

[tool.pytest.ini_options]
asyncio_mode = "auto"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "ruff (>=0.15.9,<0.16.0)",
    "mypy (>=1.20.0,<2.0.0)"
]
